// src/application/external-services/the-movie-database/the-movie-database.dto.ts
export interface InputTMDB {
    query: string[];
    include_adult?: boolean;
    language?: string;
    primary_release_year?: string;
    page?: number;
    region?: string;
    year?: string;
}

export interface OutputTMDB {
    page: number;
    results: Result[];
    total_pages: number;
    total_results: number;
}

interface Result {
    adult: boolean;
    backdrop_path?: string;
    genre_ids: number[];
    id: number;
    original_language: string;
    original_title: string;
    overview: string;
    popularity: number;
    poster_path: string;
    release_date: string;
    title: string;
    video: boolean;
    vote_average: number;
    vote_count: number;
}